debug: add debug module for dpvs
